restructuring internal package
